Split import planning preloads
