+++ src/tasks/mod.rs
@@ -1,6 +1,7 @@
 mod execution;
 mod helpers;
 mod planning;
+mod preload;
 mod remote_call;
 mod resolution;
 mod types;
